requirements.txt
anthropic
python-dotenv
pdfplumber
scikit-learn
numpy
pandas
sqlalchemy
psycopg2-binary
streamlit
tqdm
